moved nutiment to float

+++ nutriment.go
@@ -11,12 +11,12 @@ type Nutriment struct {
 	SaltServing             float64 `json:"salt_serving"`
 	SaltUnit                string  `json:"salt_unit"`
 	Sugars100G              float64 `json:"sugars_100g"`
-	Sugars                  int64   `json:"sugars"`
+	Sugars                  float64 `json:"sugars"`
 	SugarsUnit              string  `json:"sugars_unit"`
-	SugarsServing           int64   `json:"sugars_serving"`
-	SugarsValue             int64   `json:"sugars_value"`
+	SugarsServing           float64 `json:"sugars_serving"`
+	SugarsValue             float64 `json:"sugars_value"`
 	Iron                    float64 `json:"iron"`
-	IronValue               int64   `json:"iron_value"`
+	IronValue               float64 `json:"iron_value"`
 	IronLabel               string  `json:"iron_label"`
 	IronUnit                string  `json:"iron_unit"`
 	Iron100G                float64 `json:"iron_100g"`
@@ -24,79 +24,79 @@ type Nutriment struct {
 	CalciumUnit             string  `json:"calcium_unit"`
 	CalciumServing          float64 `json:"calcium_serving"`
 	Calcium                 float64 `json:"calcium"`
-	CalciumValue            int64   `json:"calcium_value"`
+	CalciumValue            float64 `json:"calcium_value"`
 	CalciumLabel            string  `json:"calcium_label"`
 	Calcium100G             float64 `json:"calcium_100g"`
 	SaturatedFat100G        float64 `json:"saturated-fat_100g"`
-	SaturatedFatServing     int64   `json:"saturated-fat_serving"`
-	SaturatedFat            int64   `json:"saturated-fat"`
-	SaturatedFatValue       int64   `json:"saturated-fat_value"`
+	SaturatedFatServing     float64 `json:"saturated-fat_serving"`
+	SaturatedFat            float64 `json:"saturated-fat"`
+	SaturatedFatValue       float64 `json:"saturated-fat_value"`
 	SaturatedFatUnit        string  `json:"saturated-fat_unit"`
 	Fat100G                 float64 `json:"fat_100g"`
-	FatServing              int64   `json:"fat_serving"`
-	FatValue                int64   `json:"fat_value"`
+	FatServing              float64 `json:"fat_serving"`
+	FatValue                float64 `json:"fat_value"`
 	FatUnit                 string  `json:"fat_unit"`
-	Fat                     int64   `json:"fat"`
+	Fat                     float64 `json:"fat"`
 	TransFatLabel           string  `json:"trans-fat_label"`
 	TransFatUnit            string  `json:"trans-fat_unit"`
-	TransFat                int64   `json:"trans-fat"`
-	TransFat100G            int64   `json:"trans-fat_100g"`
-	TransFatServing         int64   `json:"trans-fat_serving"`
-	TransFatValue           int64   `json:"trans-fat_value"`
-	VitaminA                int64   `json:"vitamin-a"`
-	VitaminA100G            int64   `json:"vitamin-a_100g"`
-	VitaminAValue           int64   `json:"vitamin-a_value"`
-	VitaminAServing         int64   `json:"vitamin-a_serving"`
+	TransFat                float64 `json:"trans-fat"`
+	TransFat100G            float64 `json:"trans-fat_100g"`
+	TransFatServing         float64 `json:"trans-fat_serving"`
+	TransFatValue           float64 `json:"trans-fat_value"`
+	VitaminA                float64 `json:"vitamin-a"`
+	VitaminA100G            float64 `json:"vitamin-a_100g"`
+	VitaminAValue           float64 `json:"vitamin-a_value"`
+	VitaminAServing         float64 `json:"vitamin-a_serving"`
 	VitaminAUnit            string  `json:"vitamin-a_unit"`
 	VitaminALabel           string  `json:"vitamin-a_label"`
-	VitaminCValue           int64   `json:"vitamin-c_value"`
+	VitaminCValue           float64 `json:"vitamin-c_value"`
 	VitaminCUnit            string  `json:"vitamin-c_unit"`
-	VitaminC100G            int64   `json:"vitamin-c_100g"`
-	VitaminC                int64   `json:"vitamin-c"`
-	VitaminCServing         int64   `json:"vitamin-c_serving"`
+	VitaminC100G            float64 `json:"vitamin-c_100g"`
+	VitaminC                float64 `json:"vitamin-c"`
+	VitaminCServing         float64 `json:"vitamin-c_serving"`
 	VitaminCLabel           string  `json:"vitamin-c_label"`
 	Proteins100G            float64 `json:"proteins_100g"`
-	ProteinsServing         int64   `json:"proteins_serving"`
-	ProteinsValue           int64   `json:"proteins_value"`
+	ProteinsServing         float64 `json:"proteins_serving"`
+	ProteinsValue           float64 `json:"proteins_value"`
 	ProteinsUnit            string  `json:"proteins_unit"`
-	Proteins                int64   `json:"proteins"`
+	Proteins                float64 `json:"proteins"`
 	Sodium                  float64 `json:"sodium"`
 	SodiumServing           float64 `json:"sodium_serving"`
 	SodiumValue             float64 `json:"sodium_value"`
 	Sodium100G              float64 `json:"sodium_100g"`
 	SodiumUnit              string  `json:"sodium_unit"`
 	CarbohydratesUnit       string  `json:"carbohydrates_unit"`
-	CarbohydratesValue      int64   `json:"carbohydrates_value"`
+	CarbohydratesValue      float64 `json:"carbohydrates_value"`
 	Carbohydrates100G       float64 `json:"carbohydrates_100g"`
-	Carbohydrates           int64   `json:"carbohydrates"`
-	CarbohydratesServing    int64   `json:"carbohydrates_serving"`
-	AlcoholValue            int64   `json:"alcohol_value"`
-	AlcoholServing          int64   `json:"alcohol_serving"`
+	Carbohydrates           float64 `json:"carbohydrates"`
+	CarbohydratesServing    float64 `json:"carbohydrates_serving"`
+	AlcoholValue            float64 `json:"alcohol_value"`
+	AlcoholServing          float64 `json:"alcohol_serving"`
 	AlcoholUnit             string  `json:"alcohol_unit"`
-	Alcohol100G             int64   `json:"alcohol_100g"`
-	Alcohol                 int64   `json:"alcohol"`
-	NovaGroup               int64   `json:"nova-group"`
-	NovaGroupServing        int64   `json:"nova-group_serving"`
-	NovaGroup100G           int64   `json:"nova-group_100g"`
-	Energy                  int64   `json:"energy"`
-	EnergyServing           int64   `json:"energy_serving"`
-	EnergyKcalServing       int64   `json:"energy-kcal_serving"`
-	EnergyKcal              int64   `json:"energy-kcal"`
-	Energy100G              int64   `json:"energy_100g"`
+	Alcohol100G             float64 `json:"alcohol_100g"`
+	Alcohol                 float64 `json:"alcohol"`
+	NovaGroup               float64 `json:"nova-group"`
+	NovaGroupServing        float64 `json:"nova-group_serving"`
+	NovaGroup100G           float64 `json:"nova-group_100g"`
+	Energy                  float64 `json:"energy"`
+	EnergyServing           float64 `json:"energy_serving"`
+	EnergyKcalServing       float64 `json:"energy-kcal_serving"`
+	EnergyKcal              float64 `json:"energy-kcal"`
+	Energy100G              float64 `json:"energy_100g"`
 	EnergyUnit              string  `json:"energy_unit"`
-	EnergyKcalValue         int64   `json:"energy-kcal_value"`
+	EnergyKcalValue         float64 `json:"energy-kcal_value"`
 	EnergyKcalUnit          string  `json:"energy-kcal_unit"`
-	EnergyKcal100G          int64   `json:"energy-kcal_100g"`
-	EnergyValue             int64   `json:"energy_value"`
-	NutritionScoreUk100G    int64   `json:"nutrition-score-uk_100g"`
-	NutritionScoreFrServing int64   `json:"nutrition-score-fr_serving"`
-	NutritionScoreFr        int64   `json:"nutrition-score-fr"`
-	NutritionScoreFr100G    int64   `json:"nutrition-score-fr_100g"`
-	NutritionScoreUkServing int64   `json:"nutrition-score-uk_serving"`
-	NutritionScoreUk        int64   `json:"nutrition-score-uk"`
-	Fiber                   int64   `json:"fiber"`
+	EnergyKcal100G          float64 `json:"energy-kcal_100g"`
+	EnergyValue             float64 `json:"energy_value"`
+	NutritionScoreUk100G    float64 `json:"nutrition-score-uk_100g"`
+	NutritionScoreFrServing float64 `json:"nutrition-score-fr_serving"`
+	NutritionScoreFr        float64 `json:"nutrition-score-fr"`
+	NutritionScoreFr100G    float64 `json:"nutrition-score-fr_100g"`
+	NutritionScoreUkServing float64 `json:"nutrition-score-uk_serving"`
+	NutritionScoreUk        float64 `json:"nutrition-score-uk"`
+	Fiber                   float64 `json:"fiber"`
 	Fiber100G               float64 `json:"fiber_100g"`
-	FiberValue              int64   `json:"fiber_value"`
-	FiberServing            int64   `json:"fiber_serving"`
+	FiberValue              float64 `json:"fiber_value"`
+	FiberServing            float64 `json:"fiber_serving"`
 	FiberUnit               string  `json:"fiber_unit"`
 }
